+++ randovania/games/samus_returns/generator/base_patches_factory.py
@@ -1,7 +1,31 @@
 from __future__ import annotations
 
+from typing import TYPE_CHECKING
+
+from randovania.games.samus_returns.layout.msr_configuration import MSRConfiguration
 from randovania.generator.base_patches_factory import BasePatchesFactory
+from randovania.generator.teleporter_distributor import (
+    get_dock_connections_assignment_for_teleporter,
+    get_teleporter_connections,
+)
+
+if TYPE_CHECKING:
+    from collections.abc import Iterable
+    from random import Random
+
+    from randovania.game_description.db.dock_node import DockNode
+    from randovania.game_description.db.node import Node
+    from randovania.game_description.game_description import GameDescription
+    from randovania.layout.base.base_configuration import BaseConfiguration
 
 
 class MSRBasePatchesFactory(BasePatchesFactory):
-    pass
+    def dock_connections_assignment(
+        self, configuration: BaseConfiguration, game: GameDescription, rng: Random
+    ) -> Iterable[tuple[DockNode, Node]]:
+        assert isinstance(configuration, MSRConfiguration)
+        teleporter_connection = get_teleporter_connections(configuration.teleporters, game, rng)
+        dock_assignment = get_dock_connections_assignment_for_teleporter(
+            configuration.teleporters, game, teleporter_connection
+        )
+        yield from dock_assignment