+++ randovania/layout/preset_migration.py
@@ -1095,6 +1095,13 @@ def _migrate_v77(preset: dict) -> dict:
     return preset
 
 
+def _migrate_v78(preset: dict) -> dict:
+    if preset["game"] == "samus_returns":
+        preset["configuration"]["teleporters"] = {"mode": "vanilla", "excluded_teleporters": [], "excluded_targets": []}
+
+    return preset
+
+
 _MIGRATIONS = [
     _migrate_v1,  # v1.1.1-247-gaf9e4a69
     _migrate_v2,  # v1.2.2-71-g0fbabe91
@@ -1173,6 +1180,7 @@ _MIGRATIONS = [
     _migrate_v75,
     _migrate_v76,  # msr door lock rando
     _migrate_v77,
+    _migrate_v78,  # msr elevator rando
 ]
 CURRENT_VERSION = migration_lib.get_version(_MIGRATIONS)